requirements.txt
langchain
langchain-core
langchain-openai
langchainhub
python-dotenv
requests
pypdf
langchain_community
faiss-gpu